requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
google-generativeai
python-multipart
aiofiles

# Optional / extras:
# openai
# httpx
# langchain
# markdown2
# pygments
